Asteroid modal + refactoring UI components
Replaces the testhome page with a new test-modal page for displaying asteroid details in a modal. Refactoring the About page's accordion usage and improves styling. Updates the accordion component for better class handling and styling. Enhances the asteroidModal component with detailed asteroid info, images, and event tables. Adds new CosmoCoin image assets and updates the navbar to link to the root path.